+++ vancouver/include/driver/logging.h
@@ -20,7 +20,7 @@
 #include "driver/vprintf.h"
 class Logging : public Vprintf
 {
-  static void (*_putcf)(void *data, long int value);
+  static void (*_putcf)(void *data, int value);
   static void *_data;
  public:
   // debug functions
@@ -28,5 +28,5 @@ class Logging : public Vprintf
   static void printf(const char *format, ...) __attribute__ ((format(printf, 1, 2)));
   static void vprintf(const char *format, va_list &ap);
   static void hexdump(const void *p, unsigned len);
-  static void init(void (*putcf)(void *, long int), void *data) { _putcf = putcf; _data = data; }
+  static void init(void (*putcf)(void *, int), void *data) { _putcf = putcf; _data = data; }
 };
